+++ db/tasks.php
@@ -34,4 +34,13 @@ $tasks = [
         'month' => '*',
         'dayofweek' => '*',
     ],
+    [
+        'classname' => '\\tool_secure_s3_storage\\task\\transfer_database_backups',
+        'blocking' => 0,
+        'minute' => '7',
+        'hour' => '*',
+        'day' => '*',
+        'month' => '*',
+        'dayofweek' => '*',
+    ],
 ];